stock_request: fix allocation and qty with backorder Currently, the modification of the allocation lines takes place after the creation of the back order. Therefore, the quantities already allocated are not correct on the backorder allocations. It is not necessary to copy an allocation that has no more open quantities. The quantity on the new allocation must also be updated with the open quantity.
